汎用要素を外部ファイルへ切り出し
