+++ src/simulations/meta_generative_art_v2/functions.ts
@@ -1,0 +1,7 @@
+import { constants } from "./constants"
+
+export function log(message: string) {
+  if (constants.system.debug) {
+    console.log(message)
+  }
+}
